function/requirements.txt
functions-framework==3.10.1
google-cloud-storage==3.10.1
google-cloud-pubsub==2.36.0
google-cloud-workflows==1.20.0